feat(input-chart): renamed category-checkbox component to categorySwitch
