+++ client/components/category-switch.js
