+++ client/components/index.js
@@ -21,4 +21,4 @@ export {default as InputChart} from './single-input-chart'
 export {default as InputSummary} from './inputSummary'
 export {default as Prediction} from './prediction'
 export {default as Recommendation} from './recommendation'
-export {default as CategoryCheckbox} from './category-checkbox'
+export {default as CategorySwitch} from './category-switch'
